bitint for money amounts, add Posting enum type to reduce chance of error in inputing LedgerLine

+++ migrations/2026-05-11-053152-0000_accounts_transactions_balances_transfers/up.sql
@@ -1,6 +1,6 @@
 -- SQL dump generated using DBML (dbml.dbdiagram.io)
 -- Database: PostgreSQL
--- Generated at: 2026-05-14T10:40:41.152Z
+-- Generated at: 2026-05-15T06:35:57.957Z
 
 CREATE TYPE "account_type" AS ENUM (
   'asset',
@@ -29,8 +29,8 @@ CREATE TABLE "ledger_lines" (
   "id" SERIAL PRIMARY KEY,
   "journal_entry_id" integer NOT NULL,
   "account" integer NOT NULL,
-  "debit" integer NOT NULL DEFAULT 0,
-  "credit" integer NOT NULL DEFAULT 0,
+  "debit" bigint NOT NULL DEFAULT 0,
+  "credit" bigint NOT NULL DEFAULT 0,
   "created_at" timestamp NOT NULL DEFAULT (now()),
   CONSTRAINT "ledger_lines_valid_leg" CHECK (debit >= 0 AND credit >= 0 AND (debit > 0) != (credit > 0))
 );
@@ -43,14 +43,14 @@ CREATE TABLE "journal_entries" (
 
 CREATE TABLE "balances" (
   "account_id" integer PRIMARY KEY,
-  "balance" integer NOT NULL DEFAULT 0,
+  "balance" bigint NOT NULL DEFAULT 0,
   "updated_at" timestamp NOT NULL DEFAULT (now())
 );
 
 CREATE TABLE "account_blocks" (
   "id" SERIAL PRIMARY KEY,
   "account_id" integer NOT NULL,
-  "amount" integer NOT NULL,
+  "amount" bigint NOT NULL,
   "created_at" timestamp NOT NULL DEFAULT (now()),
   "transfer_id" integer UNIQUE NOT NULL
 );
@@ -60,7 +60,7 @@ CREATE TABLE "transfer_internal" (
   "journal_entry_id" integer,
   "from_account_id" integer NOT NULL,
   "to_account_id" integer NOT NULL,
-  "amount" integer NOT NULL,
+  "amount" bigint NOT NULL,
   "created_at" timestamp NOT NULL DEFAULT (now()),
   "completed_at" timestamp,
   "status" transfer_status NOT NULL
